+++ src/Products/PloneMeeting/interfaces.py
@@ -576,6 +576,8 @@ class IMeetingConfigDocumentation:
     def extraAdviceTypes(self):
         '''Method for defining extra advice types, needs to return a list of
            ids that will be used for id and translated for title.'''
+    def getLateState(self):
+        '''Returns the meeting first review state from which presented items will be considered 'late'.'''
     def getMeetingStatesAcceptingItems(self):
         '''In those states, the meeting accept items, normal or late.
            Must return a tuple of meeting review_states.'''